pyproject.toml
[project]
name = "fastapi_workbench"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Javad", email = "javadsemail@gmail.com" }]
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.43",
    "asyncpg>=0.30.0",
    "alembic>=1.16.5",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib>=1.7.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "trio>=0.27.0",
    "httpx>=0.28.1",
    "black>=25.1.0",
    "ruff>=0.13.0",
    "isort>=6.0.1",
    "mypy>=1.18.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "sentry-sdk>=1.4",
    "pydantic-settings>=2.10.1",
    "aiosqlite>=0.21.0",
    "fastapi-pagination>=0.14.1",
    "python-multipart>=0.0.20",
    "psycopg2-binary>=2.9.11",
    "bcrypt>=4.3.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
